Support configuration changes from the shell. syntax: property       => lists the current value of the property property=value => Sets property to value examples: ems=true cycles=5000 core=full
Added "wc" as alias for writeconf and "wl" for writelang(Requested by
DosFreak)
Added support for home/end on the shell(parapente)
Config supports the setting of configuration as well:
config -set "section property=value"

This one should be used for properties that exist in more than one section.
Samplerate for example.

Changed the main shellinput loop. So it checks for a property if the entered
command isn't an internal command or an executable.


git-svn-id: file:///home/shlomif/progs/Work/dosbox-midi/svnsync-repos/code-0/dosbox/trunk@2190 8093ebf7-7ddc-4b6c-9c54-de804a57f08e